+++ ic-agent/src/agent/agent_test.rs
@@ -13,16 +13,69 @@ use candid::{Encode, Nat};
 use futures_util::FutureExt;
 use ic_certification::{Delegation, Label};
 use ic_transport_types::{NodeSignature, QueryResponse, RejectCode, RejectResponse, ReplyResponse};
+use reqwest::Client;
+use std::sync::Arc;
 use std::{collections::BTreeMap, time::Duration};
 #[cfg(all(target_family = "wasm", feature = "wasm-bindgen"))]
 use wasm_bindgen_test::wasm_bindgen_test;
 
+use crate::agent::http_transport::route_provider::{RoundRobinRouteProvider, RouteProvider};
 #[cfg(all(target_family = "wasm", feature = "wasm-bindgen"))]
 wasm_bindgen_test::wasm_bindgen_test_configure!(run_in_browser);
 
 fn make_agent(url: &str) -> Agent {
     Agent::builder()
         .with_transport(ReqwestTransport::create(url).unwrap())
+        .with_verify_query_signatures(false)
+        .build()
+        .unwrap()
+}
+
+fn make_agent_with_route_provider(
+    route_provider: Arc<dyn RouteProvider>,
+    tcp_retries: usize,
+) -> Agent {
+    let client = Client::builder()
+        .build()
+        .expect("Could not create HTTP client.");
+    Agent::builder()
+        .with_transport(
+            ReqwestTransport::create_with_client_route(route_provider, client)
+                .unwrap()
+                .with_max_tcp_errors_retries(tcp_retries),
+        )
+        .with_verify_query_signatures(false)
+        .build()
+        .unwrap()
+}
+
+#[cfg(feature = "hyper")]
+fn make_agent_with_hyper_transport_route_provider(
+    route_provider: Arc<dyn RouteProvider>,
+    tcp_retries: usize,
+) -> Agent {
+    use super::http_transport::HyperTransport;
+    use http_body_util::Full;
+    use hyper_rustls::{HttpsConnector, HttpsConnectorBuilder};
+    use hyper_util::{
+        client::legacy::{connect::HttpConnector, Client as LegacyClient},
+        rt::TokioExecutor,
+    };
+    use std::collections::VecDeque;
+
+    let connector = HttpsConnectorBuilder::new()
+        .with_webpki_roots()
+        .https_or_http()
+        .enable_http1()
+        .enable_http2()
+        .build();
+    let client: LegacyClient<HttpsConnector<HttpConnector>, Full<VecDeque<u8>>> =
+        LegacyClient::builder(TokioExecutor::new()).build(connector);
+    let transport = HyperTransport::create_with_service_route(route_provider, client)
+        .unwrap()
+        .with_max_tcp_errors_retries(tcp_retries);
+    Agent::builder()
+        .with_transport(transport)
         .with_verify_query_signatures(false)
         .build()
         .unwrap()
@@ -294,6 +347,73 @@ async fn status_okay() -> Result<(), AgentError> {
 
     assert!(result.is_ok());
 
+    Ok(())
+}
+
+#[cfg_attr(not(target_family = "wasm"), tokio::test)]
+async fn reqwest_client_status_okay_when_request_retried() -> Result<(), AgentError> {
+    let map = BTreeMap::new();
+    let response = serde_cbor::Value::Map(map);
+    let (read_mock, url) = mock(
+        "GET",
+        "/api/v2/status",
+        200,
+        serde_cbor::to_vec(&response)?,
+        Some("application/cbor"),
+    )
+    .await;
+    // Without retry request should fail.
+    let non_working_url = "http://127.0.0.1:4444";
+    let tcp_retries = 0;
+    let route_provider = RoundRobinRouteProvider::new(vec![non_working_url, &url]).unwrap();
+    let agent = make_agent_with_route_provider(Arc::new(route_provider), tcp_retries);
+    let result = agent.status().await;
+    assert!(result.is_err());
+
+    // With retry request should succeed.
+    let tcp_retries = 1;
+    let route_provider = RoundRobinRouteProvider::new(vec![non_working_url, &url]).unwrap();
+    let agent = make_agent_with_route_provider(Arc::new(route_provider), tcp_retries);
+    let result = agent.status().await;
+
+    assert_mock(read_mock).await;
+
+    assert!(result.is_ok());
+    Ok(())
+}
+
+#[cfg_attr(not(target_family = "wasm"), tokio::test)]
+#[cfg(feature = "hyper")]
+async fn hyper_client_status_okay_when_request_retried() -> Result<(), AgentError> {
+    let map = BTreeMap::new();
+    let response = serde_cbor::Value::Map(map);
+    let (read_mock, url) = mock(
+        "GET",
+        "/api/v2/status",
+        200,
+        serde_cbor::to_vec(&response)?,
+        Some("application/cbor"),
+    )
+    .await;
+    // Without retry request should fail.
+    let non_working_url = "http://127.0.0.1:4444";
+    let tcp_retries = 0;
+    let route_provider = RoundRobinRouteProvider::new(vec![non_working_url, &url]).unwrap();
+    let agent =
+        make_agent_with_hyper_transport_route_provider(Arc::new(route_provider), tcp_retries);
+    let result = agent.status().await;
+    assert!(result.is_err());
+
+    // With retry request should succeed.
+    let tcp_retries = 1;
+    let route_provider = RoundRobinRouteProvider::new(vec![non_working_url, &url]).unwrap();
+    let agent =
+        make_agent_with_hyper_transport_route_provider(Arc::new(route_provider), tcp_retries);
+    let result = agent.status().await;
+
+    assert_mock(read_mock).await;
+
+    assert!(result.is_ok());
     Ok(())
 }