+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+* Introduced transparent http request retry logic for network-related failures. `ReqwestTransport::with_max_tcp_errors_retries()`, `HyperTransport::with_max_tcp_errors_retries()`.
 
 * Changed the SyncCall and AsyncCall traits to use an associated type for their output instead of a generic parameter.
 * Call builders now generally implement `IntoFuture`, allowing `.call_and_wait().await` to be shortened to `.await`.